backend/requirements.txt
anyio==4.13.0
asgiref==3.11.1
attrs==26.1.0
certifi==2026.4.22
charset-normalizer==3.4.7
construct==2.10.68
construct-typing==0.5.6
dj-database-url==3.1.2
Django==6.0.4
django-cors-headers==4.9.0
djangorestframework==3.17.1
djangorestframework_simplejwt==5.5.1
drf-spectacular==0.29.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.13
inflection==0.5.1
jsonalias==0.1.1
jsonschema==4.26.0
jsonschema-specifications==2025.9.1
psycopg2==2.9.12
PyJWT==2.12.1
python-dotenv==1.2.2
PyYAML==6.0.3
referencing==0.37.0
requests==2.32.3
rpds-py==0.30.0
solana==0.34.3
solders==0.21.0
sqlparse==0.5.5
typing_extensions==4.15.0
uritemplate==4.2.0
urllib3==2.6.3
websockets==11.0.3
